+++ tests/integration/run_all.sh
@@ -116,6 +116,7 @@ ALL_CASES=(
     test_smp_init_order
     test_fpu_smp
     test_siginfo
+    test_auxv
     test_gui
     test_gui_usb
     test_gbrowser